basic jwt authentication

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,38 +1,70 @@
 import bcrypt from 'bcryptjs'
+import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
+    async login(parent, args, { prisma }, info) {
+        const user = await prisma.query.user({
+            where: {
+                email: args.data.email
+            }
+        })
+
+        if(!user) {
+            throw new Error('Unable to login')
+        }
+
+        const isMatch = await bcrypt.compare(args.data.password, user.password)
+
+        if(!isMatch) {
+            throw new Error('Unable to login')
+        }
+
+        return {
+            user,
+            token: jwt.sign( { userId: user.id }, 'thisisasecret')
+        }
+    },
     async createUser(parent,args,{prisma},info) {
         if(args.data.password.length < 8) {
             throw new Error('Password must be 8 characters or longer')
         }
 
         const password =  await bcrypt.hash(args.data.password, 10)
-
-        return prisma.mutation.createUser({
+        const user = await prisma.mutation.createUser({
             data: {
                 ...args.data,
                 password: password
             }
-        }, info)
+        })
+
+        return {
+            user,
+            token: jwt.sign({userId: user.id}, 'thisisasecret')
+        }
     },
-    async deleteUser(parent,args,{prisma},info) {
+    async deleteUser(parent,args,{prisma, request},info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
-    async updateUser(parent,args,{prisma},info) {
-        
+    async updateUser(parent,args,{prisma, request},info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
     },
-    createPost(parent,args,{prisma},info) {
-        
+    createPost(parent,args,{prisma, request},info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createPost({
             data: {
                 title: args.data.title,
@@ -40,22 +72,44 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }
         },info)
     },
-    deletePost(parent,args,{prisma},info) {
-        
+    async deletePost(parent,args,{prisma, request},info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({
             where:{
                 id: args.id
             }
         },info)
     },
-    updatePost(parent,args,{prisma},info) {
-        
+    async updatePost(parent,args,{prisma , request},info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to update post')
+        }
+
         return prisma.mutation.updatePost({
             where:{
                 id: args.id
@@ -63,13 +117,15 @@ const Mutation = {
             data: args.data
         },info)
     },
-    createComment(parent,args,{prisma},info) {
+    createComment(parent,args,{prisma, request},info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createComment({
             data:{
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -80,14 +136,38 @@ const Mutation = {
             }
         },info)
     },
-    deleteComment(parent,args,{prisma},info) {
+    async deleteComment(parent,args,{prisma, request},info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         },info)
     },
-    updateComment(parent,args,{prisma},info) {
+    async updateComment(parent,args,{ prisma, request },info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
